+++ Assets/Scripts/TransferingInputsToMovementManager/ITransferInputToMovement.cs
@@ -5,7 +5,7 @@ namespace Assets.Scripts.TransferingInputsToMovementManager
 {
     public interface ITransferInputToMovement
     {
-        public void TransferInputsToMovementData(int direction, bool isRotationalMovement);
+        public void TransferInputsToMovementData(float direction, bool isRotationalMovement);
         public void TransferInputsToMovementData(Vector3 frontDirection, Vector3 sideDirection, Vector3 rotationDirection);
     }
 }
